reajustes

+++ src/main/kotlin/com/zoop/bazoop/business/Contadigital.kt
@@ -1,5 +1,6 @@
 package com.zoop.bazoop.business
 
+import com.zoop.bazoop.application.model.ContaTO
 import model.Cliente
 //import org.graalvm.compiler.debug.TTY.print
 import java.util.Random
@@ -8,10 +9,11 @@ open class Contadigital(
     var cliente: Cliente,
     var id: Int = Random().nextInt(100000000),
     var Numconta: Int = Random().nextInt(1000) + 1,
-    var status: Status =Status.ABERTA,
+    var status: Status = Status.ABERTA,
     var tipo: Tipo = Tipo.CONTACOMUN,
-    var Saldo: Int = 1000,
-    var debito:Int = 50) {
+    var Saldo: Double = 1000.0,
+    var debito: Int = 50
+) {
     //gerar numeros automatico
     
     enum class Status
